+++ src/models/backup_file.rs
@@ -11,7 +11,7 @@ pub struct BackupFile {
     pub hash: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub enum SourceType {
     Firefox,
     Folder,